refactor(cocache): update dependencies and remove experimental stdlib opt-in
- Change kotlin("reflect") from implementation to api in cocache-core
- Remove @OptIn(ExperimentalStdlibApi::class) annotations
- Update import from kotlin.reflect to kotlin.reflect.jvm
