+++ cocache-spring/src/main/kotlin/me/ahoo/cache/spring/client/SpringClientSideCacheFactory.kt
@@ -20,7 +20,7 @@ import me.ahoo.cache.client.DefaultClientSideCacheFactory
 import me.ahoo.cache.spring.AbstractCacheFactory
 import org.springframework.beans.factory.BeanFactory
 import org.springframework.core.ResolvableType
-import kotlin.reflect.javaType
+import kotlin.reflect.jvm.javaType
 
 class SpringClientSideCacheFactory(beanFactory: BeanFactory) : ClientSideCacheFactory,
     AbstractCacheFactory(beanFactory) {
@@ -30,7 +30,6 @@ class SpringClientSideCacheFactory(beanFactory: BeanFactory) : ClientSideCacheFa
 
     override val suffix: String = CLIENT_SIDE_CACHE_SUFFIX
 
-    @OptIn(ExperimentalStdlibApi::class)
     override fun getBeanType(cacheMetadata: CoCacheMetadata): ResolvableType {
         return ResolvableType.forClassWithGenerics(
             ClientSideCache::class.java,
